Refactor file names: `quit_attempt` --> `quit_attempt_schemas`
