+++ sennight-fastapi/app/schemas/motivation_schemas.py
