+++ sennight-fastapi/app/schemas/quit_attempt_schemas.py
