+++ sennight-fastapi/app/schemas/trigger_schemas.py
